--- components/Buttons/BookNow/BookBtnDark.tsx
@@ -1,9 +1,0 @@
-import { View, Text } from 'react-native'
-const BookBtnDark = () => {
-  return (
-    <View>
-      <Text>BookBtnDark</Text>
-    </View>
-  )
-}
-export default BookBtnDark
