--- navigation/AppNavigator.tsx
@@ -1,25 +1,0 @@
-import React from 'react'
-import { createStackNavigator } from '@react-navigation/stack'
-import HomeScreen from '../screens/Home'
-import ServicesScreen from '../screens/Services'
-
-const Stack = createStackNavigator()
-
-const AppNavigator = () => {
-  return (
-    <Stack.Navigator initialRouteName="Home">
-      <Stack.Screen
-        name="Home"
-        component={HomeScreen}
-        options={{ headerShown: false }}
-      />
-      <Stack.Screen
-        name="Services"
-        component={ServicesScreen}
-        options={{ headerShown: false }}
-      />
-    </Stack.Navigator>
-  )
-}
-
-export default AppNavigator
