+++ package.json
@@ -10,16 +10,10 @@
     "ts:check": "tsc"
   },
   "dependencies": {
-    "@react-navigation/native": "^6.1.14",
-    "@react-navigation/stack": "^6.3.25",
     "expo": "~50.0.8",
-    "expo-dev-client": "^3.3.9",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
-    "react-native": "0.73.4",
-    "react-native-web": "~0.19.6",
-    "react-dom": "18.2.0",
-    "@expo/metro-runtime": "~3.1.3"
+    "react-native": "0.73.4"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
